% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sgRNA.R
\name{check_fastq_files}
\alias{check_fastq_files}
\title{Check for expected FASTQ files in a directory}
\usage{
check_fastq_files(
  timepoints = c("T0", "T4", "T10", "T14"),
  treatments = c("DMSO", "FGF401"),
  fastq_dir = NULL
)
}
\arguments{
\item{timepoints}{A character vector of timepoints to check
(default is c("T0","T4","T10","T14")).}

\item{treatments}{A character vector of treatments to check
(default is c("DMSO","FGF401")).}

\item{fastq_dir}{A character string specifying the directory
containing the FASTQ files.}
}
\value{
None. The function stops execution if any expected
FASTQ files are missing.
}
\description{
This function checks if all expected FASTQ files are present
in the specified directory based on provided timepoints and treatments.
}
