% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sgRNA.R
\name{check_sgRNA_library}
\alias{check_sgRNA_library}
\title{Check that an sgRNA library name is supported}
\usage{
check_sgRNA_library(sgRNA_library)
}
\arguments{
\item{sgRNA_library}{A character string specifying the sgRNA library}
}
\description{
Check that an sgRNA library name is supported
}
\keyword{internal}
