% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisprFlow-package.R
\docType{package}
\name{crisprFlow-package}
\alias{crisprFlow}
\alias{crisprFlow-package}
\title{crisprFlow: Functions to analyze and visualize CRISPR perturbation data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sigven/crisprFlow}
  \item \url{https://sigven.github.io/crisprFlow/}
  \item Report bugs at \url{https://github.com/sigven/crisprFlow/issues}
}

}
\author{
\strong{Maintainer}: Sigve Nakken \email{sigven@ifi.uio.no} (\href{https://orcid.org/0000-0001-8468-2050}{ORCID})

Authors:
\itemize{
  \item Sigve Nakken \email{sigven@ifi.uio.no} (\href{https://orcid.org/0000-0001-8468-2050}{ORCID})
}

}
\keyword{internal}
