% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_sgRNA_gini_index}
\alias{plot_sgRNA_gini_index}
\title{Plot sgRNA Gini index bar plot}
\usage{
plot_sgRNA_gini_index(sgRNA_summary, ggsci_style = "JCO", plot_fontsize = 11)
}
\arguments{
\item{sgRNA_summary}{A data frame containing sgRNA summary statistics including Gini index}

\item{ggsci_style}{The style for the color palette from the ggsci
package (default is "JCO")}

\item{plot_fontsize}{Font size for the plot text elements (default is 11}
}
\value{
A ggplot2 object representing the Gini index bar plot
}
\description{
This function generates a bar plot of the Gini index for sgRNA distributions
}
