% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_sgRNA_missing}
\alias{plot_sgRNA_missing}
\title{Plot sgRNA zero counts bar plot}
\usage{
plot_sgRNA_missing(sgRNA_summary, ggsci_style = "JCO", plot_fontsize = 12)
}
\arguments{
\item{sgRNA_summary}{A data frame containing sgRNA summary statistics
including zero counts}

\item{ggsci_style}{The style for the color palette from the ggsci
package (default is "JCO")}

\item{plot_fontsize}{Font size for the plot text elements (default is 12}
}
\value{
A ggplot2 object representing the sgRNA zero counts bar plot
}
\description{
This function generates a bar plot showing the number of sgRNAs with zero counts
}
