% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_sgRNA_pca}
\alias{plot_sgRNA_pca}
\title{Plot PCA of sgRNA count data}
\usage{
plot_sgRNA_pca(
  title = "Title",
  sgRNA_counts,
  ggsci_style = "JCO",
  treatment_name = "FGF401",
  plot_fontsize = 11
)
}
\arguments{
\item{title}{The title of the PCA plot (default is "Title")}

\item{sgRNA_counts}{A data frame containing sgRNA counts with sgRNA IDs as row names}

\item{ggsci_style}{The style for the color palette from the ggsci
package (default is "JCO")}

\item{treatment_name}{The name of the treatment used in the experiment (default is "
FGF401")}

\item{plot_fontsize}{Font size for the plot text elements (default is 11}
}
\value{
A ggplot2 object representing the PCA scatter plot
}
\description{
This function performs PCA on sgRNA count data and generates a scatter plot of the first two principal components.
}
