% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_plot_guide_detection}
\alias{qc_plot_guide_detection}
\title{Plot detected and lost sgRNAs per stage}
\usage{
qc_plot_guide_detection(rep_data, base_size = 12)
}
\arguments{
\item{rep_data}{Output of \code{\link{qc_guide_representation}}.}

\item{base_size}{Base font size of the plot (default is 12).}
}
\value{
A ggplot object.
}
\description{
Stacked bars of detected vs. absent/low targeting sgRNAs at each stage.
}
\examples{
rep_data <- qc_guide_representation(
  exampleCounts,
  stages = list(Plasmid = "Plasmid", T0 = c("T0_1", "T0_2", "T0_3")))
qc_plot_guide_detection(rep_data)

}
