% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{read_sgRNA_counts}
\alias{read_sgRNA_counts}
\title{Read sgRNA count data from a MAGeCK/guide-counter count file}
\usage{
read_sgRNA_counts(sgRNA_count_file)
}
\arguments{
\item{sgRNA_count_file}{The path to the MAGeCK/guide-counter count file}
}
\value{
A data frame containing sgRNA counts with sgRNA IDs as row names
}
\description{
This function reads sgRNA count data from a specified MAGeCK/guide-counter
count file, filters out non-targeting controls, and returns a data frame
}
