% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{read_sgRNA_counts2}
\alias{read_sgRNA_counts2}
\title{Read sgRNA count data from a MAGeCK/guide-counter count file
and separate targeted sgRNAs and controls}
\usage{
read_sgRNA_counts2(sgRNA_count_file = NULL, normalize = T)
}
\arguments{
\item{sgRNA_count_file}{The path to the MAGeCK/guide-counter count file}

\item{normalize}{Logical indicating whether to normalize sgRNA counts
to counts per million (default is TRUE)}
}
\value{
A data frame containing sgRNA counts with sgRNA IDs as row names
}
\description{
Read sgRNA count data from a MAGeCK/guide-counter count file
and separate targeted sgRNAs and controls
}
