% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_detection.R
\name{detect_hits_drugz}
\alias{detect_hits_drugz}
\title{Run DrugZ for hit detection}
\usage{
detect_hits_drugz(
  samplesheet_csv = NULL,
  count_table_txt = NULL,
  pseudo_count = 5,
  output_dir = NULL,
  output_prefix = "FGF401",
  drugz_python = "/Users/sigven/miniconda3/envs/drugz/bin/python"
)
}
\arguments{
\item{samplesheet_csv}{Path to the sample sheet CSV file}

\item{count_table_txt}{Path to the count table text file (from MAGeCK/guidecounter)}

\item{pseudo_count}{Pseudo count to add (default is 5)}

\item{output_dir}{Directory where output files will be saved}

\item{output_prefix}{Prefix for output files (default is "FGF401")}

\item{drugz_python}{Path to the DrugZ python interpreter (Conda environment)}
}
\value{
A list containing DrugZ hit detection results
}
\description{
This function is a placeholder for running DrugZ for hit detection.
}
