% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_detection.R
\name{detect_hits_mageck_mle}
\alias{detect_hits_mageck_mle}
\title{Run MAGeCK MLE for hit detection}
\usage{
detect_hits_mageck_mle(
  samplesheet_csv = NULL,
  count_table_txt = NULL,
  output_dir = NULL,
  norm_method = c("median", "total", "control"),
  mageck_bin = "/Users/sigven/miniconda3/bin/mageck",
  threads = 1,
  permutation_round = 4,
  late_time_start = 10,
  adjust_method = "fdr",
  output_prefix = "mageck_mle_hits",
  control_pattern = "TARGETING_NEG_CTRL"
)
}
\arguments{
\item{samplesheet_csv}{Path to the sample sheet CSV file}

\item{count_table_txt}{Path to the count table text file (from MAGeCK
count or guidecounter)}

\item{output_dir}{Directory where output files will be saved}

\item{norm_method}{Normalization method to use ("median", "total", or "
control")}

\item{mageck_bin}{Path to the MAGeCK binary}

\item{threads}{Number of threads to use (default is 1)}

\item{permutation_round}{Number of permutation rounds (default is 2, recommended 10 (longer time))}

\item{late_time_start}{Time point to start considering as "late" (default is
10)}

\item{adjust_method}{Method for p-value adjustment (default is "fdr")}

\item{output_prefix}{Prefix for output files (default is "crisprFlowRun")}

\item{control_pattern}{Pattern to identify control genes (default is "TARGETING_NEG_CTRL")}
}
\value{
A data frame containing the gene summary results from MAGeCK MLE
}
\description{
This function runs MAGeCK MLE for hit detection using a provided sample sheet
and count table. It constructs a design matrix based on the sample information
and executes the MAGeCK MLE command with specified parameters.
The results are saved in the specified output directory.
}
