% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{read_sgRNA_summary}
\alias{read_sgRNA_summary}
\title{Read sgRNA summary statistics from a MAGeCK/guide-counter summary file}
\usage{
read_sgRNA_summary(
  sgRNA_count_summary_file = NULL,
  timecourse_drug_screen = TRUE
)
}
\arguments{
\item{sgRNA_count_summary_file}{The path to the MAGeCK/guide-counter count
summary file}

\item{timecourse_drug_screen}{logical indicating whether the data is from a timecourse drug screen (default is TRUE)}
}
\value{
A data frame containing cleaned sgRNA summary statistics
}
\description{
This function reads sgRNA summary statistics from a specified
MAGeCK/guide-counter summary file, cleans the column names,
and returns a data frame
}
